move(cosmo): move core to private API
